+++ crates/matrix-sdk-crypto/CHANGELOG.md
@@ -54,6 +54,9 @@ Deprecations:
 
 Additions:
 
+- Expose new method `OlmMachine::upload_device_keys()`.
+  ([#3457](https://github.com/matrix-org/matrix-rust-sdk/pull/3457))
+
 - Expose new method `CryptoStore::import_room_keys`.
   ([#3448](https://github.com/matrix-org/matrix-rust-sdk/pull/3448))